Binary CAN sniffer
This sketch is far more efficient than the plaintext variant.
It writes the CAN message data to the SD card in binary and flushes only as often as necessary.
Additionally, the RTC's 1 Hz square wave generator is used to store only the number of seconds since start, which reduces the number of queries for the RTC to just once per trip.